pyproject.toml
[project]
name = "myscc26"
version = "0.1.1"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "argon2-cffi>=25.1.0",
    "bcrypt>=5.0.0",
    "black>=26.3.1",
    "dj-rest-auth>=7.2.0",
    "django==6",
    "django-allauth>=65.15.0",
    "django-cors-headers>=4.9.0",
    "django-debug-toolbar>=6.2.0",
    "django-sslserver>=0.22",
    "djangorestframework>=3.17.0",
    "djangorestframework-simplejwt>=5.5.1",
    "pillow>=12.1.1",
    "pip>=26.0.1",
    "reportlab>=4.4.10",
    "ruff>=0.15.7",
]
